* Common/BitUtils.h (CPUID): Values are 32-bit, not 64-bit.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@72201 ad02020f-2a26-4133-947e-1829317c1416

+++ Common/BitUtils.h
@@ -7,7 +7,7 @@
 #include <stdint.h>
 
 /** The return value of the CPUID instruction. */
-struct CPUID { uint64_t a, b, c, d; };
+struct CPUID { unsigned a, b, c, d; };
 
 /** Return the result of the CPUID instruction
  * or -1 if it is not supported.
